moved apple hardware test to separate file
